added logging files
